Привязка дома к участку

+++ app/modules/application/land.controller.php
@@ -69,9 +69,9 @@ class LandController extends BaseController {
             $land->number = Input::get('number');
             $land->area = Input::get('area');
             $land->price = Input::get('price');
-            $land->coordinate_x = Input::get('coordinate_x');
-            $land->coordinate_y = Input::get('coordinate_y');
-            $land->photo_id = Input::get('photo_id');
+            $land->coordinate_x = (int)Input::get('coordinate_x');
+            $land->coordinate_y = (int)Input::get('coordinate_y');
+            $land->photo_id = (int)Input::get('photo_id');
             $land->sold = 0;
             $land->save();
 
@@ -110,9 +110,9 @@ class LandController extends BaseController {
             $land->number = Input::get('number');
             $land->area = Input::get('area');
             $land->price = Input::get('price');
-            $land->coordinate_x = Input::get('coordinate_x');
-            $land->coordinate_y = Input::get('coordinate_y');
-            $land->photo_id = Input::get('photo_id');
+            $land->coordinate_x = (int)Input::get('coordinate_x');
+            $land->coordinate_y = (int)Input::get('coordinate_y');
+            $land->photo_id = (int)Input::get('photo_id');
             $land->sold = Input::has('sold') ? 1 : 0;
             $land->save();
